+++ src/components/Penalty/organisms/PenaltyTable/style.ts
@@ -2,13 +2,11 @@ import styled from '@emotion/styled';
 import { Palette } from 'styles/globals';
 
 export const TableWrapper = styled.div`
+  width: 100%;
   height: 87.5vh;
   display: flex;
   justify-content: space-between;
-  gap: 30px;
-  @media (max-width: 950px) {
-    gap: 0;
-  }
+
   @media (max-width: 420px) {
     display: flex;
     flex-direction: column;
@@ -18,7 +16,7 @@ export const TableWrapper = styled.div`
 `;
 
 export const ListWrapper = styled.div`
-  width: 80%;
+  width: 100%;
   height: 100%;
   padding: 0 16px 16px;
   overflow-y: scroll;
@@ -111,48 +109,6 @@ export const TagItem = styled.div`
   svg {
     width: 13px;
     cursor: pointer;
-  }
-`;
-
-export const CSVWrapper = styled.div`
-  min-width: 250px;
-  width: 280px;
-  color: ${Palette.NEUTRAL_N10};
-  background-color: ${Palette.BACKGROUND_CARD};
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  border-radius: 16px;
-  margin-top: 24px;
-  span {
-    font-weight: 700;
-    font-size: 18px;
-  }
-  @media (max-width: 1200px) {
-    width: 23.5vw;
-  }
-  @media (max-width: 950px) {
-    display: none;
-  }
-`;
-
-export const CSVLink = styled.div`
-  height: 45px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: ${Palette.SYSTEM_POSITIVE};
-  font-weight: 700;
-  font-size: 16px;
-  border: 1px solid ${Palette.SYSTEM_POSITIVE};
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.1s ease-in-out;
-  :hover {
-    border: none;
-    color: ${Palette.BACKGROUND_CARD};
-    background-color: ${Palette.SYSTEM_POSITIVE};
   }
 `;
